refactor : move monster selector and hooks out of GameControlsComponent
